Вторая версия кода_улучшенная

+++ src/tracker/controllers/HistoryManager.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface HistoryManager {
     public void add(Task task);
-    public void remove(int id);
+    public void remove(Long id);
     public List<Task> getHistory();
 }
